requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
google-generativeai